Don't depend on DNS for Puppet execution
Facter's idea of $::fqdn depends on there being a DNS entry for the
current host. Since Foreman may or may not do that for us, force
Puppet's idea of the host name.

Cons: this makes it harder to rename a node.

Pros: this makes it possible to provision nodes and run Puppet in case
of a partial SNAFU in the Foreman database.